Extract scan.ts in own module
